+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/model/EcompUserDetails.java
@@ -33,7 +33,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 public class EcompUserDetails implements UserDetails {
 
 	private static final long serialVersionUID = 1L;
-	private final EcompUser ecompUser;
+	// use transient per Sonar warning
+	private transient final EcompUser ecompUser;
 
 	// This is the default Spring role-name prefix.
 	private static final String ROLEP = "ROLE_";