+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/util/HttpsURLConnectionUtils.java
@@ -22,12 +22,12 @@ package org.oransc.ric.portal.dashboard.util;
 
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
+import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSession;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
@@ -42,28 +42,26 @@ public final class HttpsURLConnectionUtils {
 
 	private static final HostnameVerifier jvmHostnameVerifier = HttpsURLConnection.getDefaultHostnameVerifier();
 
-	private static final HostnameVerifier trivialHostnameVerifier = new HostnameVerifier() {
-		public boolean verify(String hostname, SSLSession sslSession) {
-			return true;
-		}
-	};
+	private static final HostnameVerifier trivialHostnameVerifier = (hostname, sslSession) -> true;
 
 	private static final TrustManager[] UNQUESTIONING_TRUST_MANAGER = new TrustManager[] { new X509TrustManager() {
 		public java.security.cert.X509Certificate[] getAcceptedIssuers() {
 			return null;
 		}
 
-		public void checkClientTrusted(X509Certificate[] certs, String authType) {
+		public void checkClientTrusted(X509Certificate[] certs, String authType) throws CertificateException {
+			// empty implementation
 		}
 
-		public void checkServerTrusted(X509Certificate[] certs, String authType) {
+		public void checkServerTrusted(X509Certificate[] certs, String authType) throws CertificateException {
+			// empty implementation
 		}
 	} };
 
 	public static void turnOffSslChecking() throws NoSuchAlgorithmException, KeyManagementException {
 		HttpsURLConnection.setDefaultHostnameVerifier(trivialHostnameVerifier);
 		// Install the all-trusting trust manager
-		SSLContext sc = SSLContext.getInstance("SSL");
+		SSLContext sc = SSLContext.getInstance("TLS");
 		sc.init(null, UNQUESTIONING_TRUST_MANAGER, null);
 		HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
 	}
@@ -71,7 +69,7 @@ public final class HttpsURLConnectionUtils {
 	public static void turnOnSslChecking() throws KeyManagementException, NoSuchAlgorithmException {
 		HttpsURLConnection.setDefaultHostnameVerifier(jvmHostnameVerifier);
 		// Return it to the initial state (discovered by reflection, now hardcoded)
-		SSLContext sc = SSLContext.getInstance("SSL");
+		SSLContext sc = SSLContext.getInstance("TLS");
 		sc.init(null, null, null);
 		HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
 	}